fix: long class on question

See merge request atixlabs-oss/semillas-middleware!148

+++ src/main/java/com/atixlabs/semillasmiddleware/excelparser/app/constants/DidiSyncStatus.java
@@ -39,13 +39,13 @@ public enum DidiSyncStatus implements CategoryQuestion {
     ANTIQUITY("¿DESDE QUE ANO VIVIS EN ESTA VIVIENDA?"){
         @Override
         public Class<?> getDataType() {
-            return Integer.class;
+            return Long.class;
         }
     },
     NUMBER_OF_ENVIRONMENTS("CANTIDAD DE AMBIENTES"){
         @Override
         public Class<?> getDataType() {
-            return Integer.class;
+            return Long.class;
         }
     },
     RENTAL("MONTO ALQUILER"){
